+++ packages/auth-cognito/CHANGELOG.md
@@ -2,6 +2,36 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.1.1-beta.11](https://github.com/ten24group/fw24-modules/compare/v0.1.1-beta.10...v0.1.1-beta.11) (2025-04-28)
+
+### [0.1.1-beta.10](https://github.com/ten24group/fw24-modules/compare/v0.1.1-beta.9...v0.1.1-beta.10) (2025-04-28)
+
+### [0.1.1-beta.9](https://github.com/ten24group/fw24-modules/compare/v0.1.1-beta.8...v0.1.1-beta.9) (2025-04-23)
+
+### [0.1.1-beta.8](https://github.com/ten24group/fw24-modules/compare/v0.1.1-beta.7...v0.1.1-beta.8) (2025-04-07)
+
+### [0.1.1-beta.7](https://github.com/ten24group/fw24-modules/compare/v0.1.1-beta.6...v0.1.1-beta.7) (2025-04-01)
+
+### [0.1.1-beta.6](https://github.com/ten24group/fw24-modules/compare/v0.1.1-beta.5...v0.1.1-beta.6) (2025-03-18)
+
+### [0.1.1-beta.5](https://github.com/ten24group/fw24-modules/compare/v0.1.1-beta.4...v0.1.1-beta.5) (2025-03-14)
+
+### [0.1.1-beta.4](https://github.com/ten24group/fw24-modules/compare/v0.1.1-beta.3...v0.1.1-beta.4) (2025-03-10)
+
+
+### Bug Fixes
+
+* package import error ([8a41167](https://github.com/ten24group/fw24-modules/commit/8a4116773ec6b452395f2b704ef28edbb8d6a2c4))
+
+### [0.1.1-beta.3](https://github.com/ten24group/fw24-modules/compare/v0.1.1-beta.2...v0.1.1-beta.3) (2025-03-03)
+
+
+### Bug Fixes
+
+* env-keys prefix for AuthModule ([2f9ac1e](https://github.com/ten24group/fw24-modules/commit/2f9ac1ea9b9f6898f209c41976afeee88f96e1b3))
+
+### [0.1.1-beta.2](https://github.com/ten24group/fw24-modules/compare/v0.1.1-beta.1...v0.1.1-beta.2) (2025-02-28)
+
 ### [0.1.1-beta.1](https://github.com/ten24group/fw24-modules/compare/v0.1.1-alpha.2...v0.1.1-beta.1) (2024-12-04)
 
 ### [0.1.1-alpha.2](https://github.com/ten24group/fw24-modules/compare/v0.1.1-alpha.1...v0.1.1-alpha.2) (2024-12-04)